Added contributing instructions
I added some instructions to the repo in a README file so that other members
of the team would know how to add code to the repo.

I aimed to make the instructions clear and simple to follow. I also wanted to
make sure that people left good context for the contributions that they were
making, so I added quite a lot about commit messages.