Document how we should document our work
Documenting code in Confluence means that people have to know to look
there, and that the documentation can't be reviewed with the code it's
meant to document. This can mean that the documentation either doesn't
get updated, is updated too soon, or isn't reviewed properly. To that
end, we should always document our work as close to the code as
possible.