Move aws_federation TF module to global resources
As the aws_federation Terraform module creates a temporary webops admin IAM role with a fixed name it currently causes an error when attempting to apply Terraform resources in multiple workspaces, as IAM role names are unique.

In the long term IAM roles should be created dynamically, either via Github events or as part of the identity broker authentication flow. For now though, the simplest thing to do is to move the federation module into the `global-resources` TF resources, as they are not workspace-specific.

As this modifies both `global-resources` and `cloud-platform` state, `terraform apply` will need to be run in the `global-resources` context, and each `cloud-platform` workspace.