Increase live-0 worker node count
Due to the nature of EBS volumes being tied to a single AZ in AWS, when performing a rolling-update in nodes there is a
period between terminating a node and when the replacement is launched during which, pods using EBS volumes are unable
to schedule.

This change doubles the number of nodes so that there is always at least one available per zone during a rolling update.