Re-introduce the HTTPS redirection server snippet
Slightly modified to use $scheme instead of $http_x_forwarded_proto (which is not available anymore since we do L4 load
balancing). Unfortunately, even though `ssl-redirect` is enable by default and redirects http to https, it does not do
this for the default backend. We cannot use the `force-ssl-redirect` annotation since the default backend does not use
an Ingress. Therefore, the only way to have the default backend be redirected to https at the moment is to simply add
this block back in.