Update ingress conflict policy
Use := for assignment and == for equality. This makes the policy easier to read and is recommended practice. The
previous policy would fail to run properly in the opa repl.

Additionally, it removes the operation type check (redundant, we want to check both CREATE and UPDATE) which is set at
the ValidatingWebhookConfig level. Also removes the namespace check (which invalidates the check for ingresses within
the same namespace).