Parse nginx-ingress log records as JSON
This filters any logs coming from an nginx-ingress container (based on the file name, this is not a perfect check and
we can probably do some re-tagging based on metadata to have a more robust implementation) and tries to parse them as
JSON or else treat them as a string.

The logs are sent to separate indices on elasticsearch (logstash-nginx-ingress-<date>) to allow for proper mappings to
be inferred.