Update fluentd configuration from upstream
There are a lot of changes due to the new formatting.
Since our configuration has slightly shifted from upstream, this is the merge of the two.

The final section of `containers.input.conf` has been intentially left commented out while we investigate the best way
forward, when it comes to parsing structured application logs.