Remove PriorityClass from fluentd chart
This was added to the helm chart to solve a dependency issue but it's a cluster resource which is not specific to
fluentd. The set of two PriorityClasses were supposed to be managed through the environments repository but this does
not work well for new clusters. Therefore, the proper solution is to introduce them as cluster components and define
the dependency of fluentd on them.