Modify chunk_limit_size to 100MB
Increase the number of chunks to maintain the same total buffer space.

This is because the plugin will try to send entire chunks if they're full and elasticsearch has a limit,
`http.max_content_length` on the request size. This limit is 100MB by default and is not configurable on AWS ES.

See https://github.com/uken/fluent-plugin-elasticsearch#buffer-options for details.