Set default_elasticsearch_version and type_name
In elasticsearch version 7, type_name is deprecated and the only valid value is `_doc`. In version 6, type_name is
configurable but we can only have a single value in an index. Ideally, we'd want to switch to `_doc` but due to the
limitation this presents a problem in performing the upgrade on a live cluster.

Since we do not have the ability to upgrade elasticsearch to version 7 yet, we should be explicit about this choice.