Create cluster in a sub-shell
The reason kops can't find its yaml file is that
pwd is terraform/cloud-platform when the create
cluster part is finished.
By running it in a sub-shell, the pwd should be
unchanged by the time we run kops, so the relative
directory path should be correct, and kops should
find its yaml file.