Don't run sudo commands if we are root
Inside the tools image, we are running as the
root user, and there is no `sudo` executable. So,
these commands would cause the build script to
fail.

I'm bypassing them with guards for now, but a
better option would be to change the tools image
to not use a root user.