Stop flushing DNS cache, waiting for kops validate
The cluster build process works without needing a DNS cache flush,
although it probably takes about 5 minutes longer, when running on
someone's laptop. When we can run the build in a kubernetes job,
there will probably be no difference in the run time.

Removing the DNS flush step means we no longer need sudo. This
simplifies the script, and also makes it easier to convert it to
run as a kubernetes job.