sanity check if running in container (#370)
* sanity check if running in container

* Move the docker check to the bottom of the file

Everything near the top of the file is definitions
rather than code which is executed immediately.

Moving this check to the bottom of the file keeps
things consistent

* Extract the file checks to a method

This makes the intention easier to understand

* Make the abort message more explicit

Although we're testing that we're running in any
container, we should be running inside the tools
image. This change makes this clear to the user.