Increase logging test sleep from 60 to 120 seconds
The logging spec has been failing intermittently in the concourse
pipeline. e.g.

* [build 547](https://concourse.cloud-platform.service.justice.gov.uk/teams/main/pipelines/integration-tests/jobs/apply-tests-live-1/builds/547)
* [build 533](https://concourse.cloud-platform.service.justice.gov.uk/teams/main/pipelines/integration-tests/jobs/apply-tests-live-1/builds/533)
* [build 526](https://concourse.cloud-platform.service.justice.gov.uk/teams/main/pipelines/integration-tests/jobs/apply-tests-live-1/builds/526)
* [build 517](https://concourse.cloud-platform.service.justice.gov.uk/teams/main/pipelines/integration-tests/jobs/apply-tests-live-1/builds/517)

This commit is an experiment to see if a larger sleep delay makes
these intermittent failures go away.

If it does, subsequent commits will attempt to zero in on the
lowest delay which gives us consistently passing tests.

If it doesn't, we know the problem is something else, and we can
focus efforts elsewhere to find the real issue.