Fix kiam test for new clusters
When I ran the integration tests for a new cluster
the kiam test failed.

On investigation, this was because the "trust
relationships for the relevant IAM role are
specific to a cluster.

https://console.aws.amazon.com/iam/home?region=eu-west-2#/roles/integration-test-kiam-iam-role?section=trust

i.e. whichever cluster is being tested when the
role is created, only that cluster's nodes were
allowed to assume the role.

This commit queries the list of 'principals' in
the role's trust relationships, and adds the
current cluster if its nodes are not listed.

This has to be done by tweaking and replacing the
whole policy as a JSON document, because of
limitations in the AWS API.

I tried setting a wildcard pattern in the list of
principals, to avoid having to do this for every
new cluster, but AWS rejected that as a malformed
value. It also rejects any values for non-existent
clusters.

This has the drawback that the list of principals
will grow to include an entry for every test
cluster we ever create (unless AWS is smart enough
to remove entries when the underlying cluster is
deleted).