Accomodate paginated list_roles responses
This test calls list_roles. It started failing
because we (MoJ Digital) hit 100 roles in the AWS
account, which is the maximum number of roles we
can get in the response. 

This change keeps calling list_roles until all 
defined roles have been returned, which makes the
test pass again.