Replace get_running_pod_name method
This helper method was rather hacky. This change
uses a more robust mechanism to get the name of
the running pod, and brings that into line with
all the other helper methods that interrogate the
cluster to get information about pods.

This commit also removes the (now) unused
get_pod_name method, which was similarly hacky.