Remove cluster nodes from role policy, after tests
After the kiam tests have finished, remove the
cluster's nodes from the role policy trust
relationships.

This avoids a problem where, after a cluster is
deleted, its ARN in the trust relationships list
is replaced with a malformed value (AWS does this
by itself).

Pre-emptively removing the cluster from the trust 
policy means this won't happen.

The exception is live-1. We never remove live-1
nodes from the list, which means we don't have
to worry about the list size dropping to zero, 
which would complicate things.