Add a script to cleanly destroy a (test) cluster
This script destroys a test cluster by reversing the steps used to
create it.

It also deletes the prometheus_operator helm release first, because
it's common to have problems with this if it happens concurrently
with deleting the rest of the components.
Finally, the script clears up terraform workspaces.

This script is deliberately NOT written to take a cluster name
as a command-line parameter, as an extra safety measure.