Move node machineType to terraform variables
Instead of hard-coding EC2 instance types into
the kops config template, this change moves them 
to terraform variables.

This change by itself will not alter anything, but
it's a pre-requisite for the next change to the 
ruby script which builds test clusters. That 
change will alter the invocation of the terraform
apply step, to supply different machine type
values, so that we can build different sizes of
test cluster.