Add dependency on ingress-controllers
The installation of prometheus operator fails the first time, when
we build a test cluster.

    Error: Error applying plan:

    1 error occurred:
            * helm_release.prometheus_operator: 1 error occurred:
            * helm_release.prometheus_operator: rpc error: code = Unknown
    desc = release prometheus-operator failed: Internal error occurred:
    failed calling webhook "validate.nginx.ingress.kubernetes.io": Post
    https://nginx-ingress-acme-controller-admission.ingress-controllers.svc:443/extensions/v1beta1/ingresses?timeout=30s:
    dial tcp 100.71.204.116:443: connect: connection refused

I think this is because the admission webhook is not ready when
the installation happens.

This change adds a dependency, so that terraform doesn't try to
install prometheus operator before the webhook is in place.