Bugfix: correct the command which disables alerts
We want to replace the pagerduty token and all
slack webhooks in test clusters, so we run a `sed`
command on the terraform source to replace those
values with `dummydummy`

But, because the code is executed via ruby, we
need to escape every backslash by doubling it
(i.e. wherever you put `\` in bash, e.g. when
running the command a terminal, you need to put
`\\` in your ruby code.

I think this is why we've still been getting 
alerts from our test clusters.