Add a timeout to the node drain step
Draining a node can fail for various reasons. When
this happens, the drain command loops forever (or,
at least for longer than I was prepared to wait,
during testing).

This commit adds a timeout to the node drain step,
so that we only wait a fixed amount of time before
giving up (and killing the spawned `kubectl`
process).