Configure kops to use a docker registry cache
This change adds a docker registry to the live-1 cluster:
https://github.com/ministryofjustice/cloud-platform-environments/pull/1749

This commit alters the kops config, and the template for new kops
clusters, to use a registry cache via the URL provided.

Once this change has been merged, a rolling update of the live-1
cluster will be required to get worker nodes to start using the
cache.