Correct the instancegroup name to nodes-r52xl
The last time we changed the cluster, we increased
the node size to r5.2xlarge by creating a new
instancegroup called nodes-r52xl and then deleting
the old instancegroup called "nodes"

At that time, we didn't update the yaml file in
the github repo.

This change is required so that we can resize the
worker nodes via a kops rolling-update

Clearly, this instancegroup name will be wrong
after the resize, however it's the only way to use
a rolling-update to do the resize, so we'll have
to live with it, for now.