Refactor: extract `kubectl_items` method
This change replaces several instances where we
send `kubectl get ...... -o json`, parse the
resulting JSON and then return the value of
"items".

Instead of that, this commit adds a method which
takes the command string, prepends "kubectl" and 
appends " -o json", parses the result and returns
the "items" part.

Hopefully, this is easier to read and more 
explicit than a lot of copy/pasted one-liners.