Restrict inbound access to the cache ingress
To restrict traffic so that only our cluster can
access the docker registry, we need to specify the
nat gateway public IPs of the VPC in which the 
cluster lives.

This commit is a hard-coded hack to demonstrate
the principle. A later commit will get the correct
IPs from the cluster terraform state, so that we
have the right values for new clusters.