Increase wait time for cache log output
The test which confirms that the docker registry cache is actually
being used by the cluster worker nodes is failing intermittently
on live-1.

This change increases the sleep delay between pulling an image and
checking that the cache logged some output from 1 to 5 seconds, to
try and make this test more reliable.