Update cluster build wrt. bundler
The latest version of bundler, which is installed
on the fly during the cluster build process, works
differently wrt. the executables provided by gems
(e.g. the `rspec` command provided by the `rspec-
core` gem).

This change ensures that the integration tests run
correctly at the end of the cluster build script.