Change the ID of a namespace resource
The docker-registry-cache namespace is defined in a kubernetes_namespace
resource which had the identifier "docker_registry_cache"

This was confusing the orphaned namespace checker, which reports that
the "docker-registry-cache" namespace is not defined in our source code.

Changing the identifier to match the namespace name should resolve this
problem (assuming terraform will allow an identifier with hyphens in
it).