Add more full-featured `execute` method
I need to change the helper file so that all
`system(cmd)` and backtick command invocations go
through the execute method, so that later on I can
add a delay to it, to slow down *all* kubectl
commands.

To do that, I need to ensure the execute method
can be used both when the caller wants the
*return status* and when the caller wants the
stdout text resulting from invoking the command.

Hence this change, which uses Open3 to return
stdout, stderr, and the status of the command
execution.