Remove tf.log file
This file should never have been committed to the repo. I must have
accidentally included it in the November 2019 commit.

I've checked the file, and the only problematic things I can find in
there are:

  * the access key (but not the secret key) from my AWS credentials
  * a private TLS key for accessing a test cluster
  * The oidc_components_client_secret and oidc_kubernetes_client_secret
    for the same test cluster

That test cluster has long since been destroyed.

I've invalidated those AWS credentials, just to be safe, but I don't
think there is anything we need to be worried about. The file
certainly shouldn't be there, so this commit removes it.