Log whenever the tests sleep, for the pipeline
We have started seeing KubeAPILatencyHigh errors again, although
they seemed to have more or less stopped after we added this delay.
They started again after we moved the integration test pipeline to
the manager cluster concourse instance.
This commit is a temporary change to 100% confirm that the sleep
delay is still happening, in the new pipeline.