Remove the sleep delay when running via pipeline
We have disabled the error page integration tests, because we
think they're responsible for the "KubeAPILatencyHigh" alerts we
often receive when the tests are running.
As such, this additional sleep delay should not be necessary, so
this commit removes it.
An additional commit will remove the environment variable from the
integration tests concourse pipeline.