Revert "Remove sleep when running via pipeline"
This reverts commit 4d0bc44c5f24d64eb7c43c9446f6cf4a9b2cc836.

requires
https://github.com/ministryofjustice/cloud-platform-concourse/pull/194

It seems that the integration tests send a lot of AWS API calls,
particularly to the route53, where we have an API requests/second
limit (of 7, a global limit we can't alter).

This change reinstates the sleep delay between every command that
the tests execute, to reduce the chances of the integration tests
pushing us over that limit again.