Temporarily remove eks.tf
This file has been added to .gitattributes by commit e75f22ffc
But, since the file was already in the repo in cleartext, this
breaks git-crypt for everyone. The effect is that your working
copy is never clean, because `terraform/cloud-platform-eks/eks.tf`
always shows as modified, even if you try to use `git checkout` to
go back to the unmodified version.

I'm hoping that using git remove and then a second PR to git
add the encrypted version will fix this.