Add common kube-system service accounts to privileged
EKS requires a number of service accounts in kube-system to run as root
contains or with the ability to escalate privileges. After deleting the
default PSP (eks.privileged) we require a way for these service accounts
to assume the relevant permissions.

As we already have a privileged pod PSP, it seems sensible to add the
necessary service accounts to our role privileged role  binding,
allowing them the permission required.