Remove EKS specific label from PSP
As we will be using this policy in multiple clusters (kops and EKS),
this label is unnecessary.

According to their user guide, the label acts as a preventive measure
against installing multiple critical components, such as DNS. Our PSP's
work without this label so I'm happy to remove it.

Source:
https://docs.aws.amazon.com/eks/latest/userguide/eks-ug.pdf