Move begin/rescue inside the 'records' loop
I made a mistake in #833, whereby the rescue block is outside of the
context in which `record` exists, so an undefined variable error gets
thrown if the rescue clause is triggered.

This change moves the begin/rescue inside the loop, where `record`
exists if an error occurs. This should have the desired effect of
allowing the tests to continue successfully, even if the A/TXT record is
not found when the cleanup code tries to delete it.