Change master nodes from c4.4xlarge to c5.4xlarge
This is because the c4 instance family is nearing its end of life, and
we ran into availability problems the last time we tried to replace a c4
master node.

This change should help to mitigate such issues, because c5 has
generally better availability. We are talking to AWS about a more robust
way to prevent such problems in future.