Remove fluentd integration tests
Please note, this will not remove fluentd from our cluster, just the
ability to test it. We are aware of issues related to this integration
test and a decision has been made to remove fluentd altogether. The
removal will be undertaken in another ticket.

Fluentd still exists in the manager cluster so all tests that take place
there are untouched.