Add three ingress instance groups
This commit adds three additional instance groups to the Kops manifest.
These groups contain a single instance in each eu-west-2 region to allow
for fine grained region failure and scale.

The idea behind these instance groups is to dedicate a resource pool to
a critical component in a production cluster.