Add a temporary hack while ingress-nodes exist
We have nodes dedicated to running ingress
controllers, which we're not using yet.

While these nodes exist, they must be excluded
from tests which collect node IPs to see whether
pods in daemonsets are running on "all" nodes.

We don't plan on keeping the ingress nodes, so
this hack should be reverted soon.