Add tests for different ingress controllers
These specs check that known ingresses on each of
the "nginx" and "modsec01" ingress controllers
respond as expected.

In particular, the specs confirm that the "Server"
header is not returned in the HTTP response.

As of now, this is true for the "modsec01" 
controller, but not for the "nginx" controller,
which needs a small configuration change to 
suppress that header. So, this test will fail
until that change is made.

These tests are quite brittle, because they depend
on an ingress which we don't control.