Delete offender-search-dev.tf
The offender-search-dev namespace was renamed (i.e. deleted and a new namespace created) to `probation-offender-search-dev` back in July. 

Although there isn't an equivalent `probation-offender-search-dev.tf` file in this directory, I think this file should be deleted because it can't possibly be doing anything useful, since it points to a namespace that doesn't exist anymore.