Increase EBS volume for ES clusters
Both the main and the audit cluster are showing
"ClusterIndexWritesBlocked", and not recording any data.

This change should double the disk space for all cluster nodes.

This is not sufficient to fix the problem. After the disk space is
available, we also need to take the clusters out of read-only mode as
described in this blog post:

https://engineeernitesh.blogspot.com/2019/12/elasticsearch-error-clusterblockexcepti.html