Add Elasticsearch monitoring
This commit connects to
https://github.com/ministryofjustice/cloud-platform/issues/2614 and
relates to the addition of Cloudwatch monitoring on our Elasticsearch
instances: live and audit.

To make things easier, I decided to use modules to create the alerts.

The first module: terraform-aws-modules/notify-slack/aws
was created by AWS and implements a Lambda script, an SNS topic and
associates them with a Slack webhook, which is passed via
terraform.tfvars.

The second module: dubiety/elasticsearch-cloudwatch-sns-alarms/aws
us a community module which uses the sns topic (created in the module
above) and sends "best practice" alarms to Slack when their status
changes.

This makes maintaining and upgrading Elasticsearch monitoring easier and
keeps things very neat, from our point of view.

When we upgrade to Terraform 0.13+ we will need to bump the notify_slack
module version as this will no longer be supported.