Add aws-node pod security policy
The aws-node component requires the `NET_RAW` capability. This was
dropped by all resources as part of
https://github.com/ministryofjustice/cloud-platform-infrastructure/pull/943
and should remain dropped.

As the component requires it and there is no sight of when it'll be
amended, it was suggested that a separate PSP is created and maintained
supporting only `aws-node`.