Add fix_psp method to create-script
There's no way to bootstrap a cluster with your own PSP's in EKS, which
means we have to get creative. This method simply deletes the default
EKS policy, installs our own and then cycles all pods in the cluster.

It needs to be executed after the cluster is created and before we bulk
it out with components.