Amend eks module to allow for live creation
It turns out to be fairly simple, we would like to create as close to
`live-1` as possible. `Live-1` is currently running 19 worker nodes and
three masters. At the start of `Live` it will need a different instance
type `r5.xlarge` for mem optimisation but we can keep the instance count
the same and utilise the out of the box autoscaling.

There are also a few tags that need to be included as part of our
tagging strategy.

I wanted to keep the if conditions down as much as possible. This can be
iterated in the future.