Increase waiting time after creating ingress (#1239)
The  cert-manager spec is consistently failing because it failed to resolve the DNS for host created. Even though the ingress was created in <90s and the external DNS entry created in A record around 114s, It takes some time to get the DNS resolved.

When ran locally, it passed when the wait time is 240. It did failed sometimes when set to 180.