refactor(infrastructure testing): :construction: refactor certificate check to utilise before and after specs
Ginkgo doesn't provide an option for before a context or describe, only a before "it" or before suite. I believe this feature is coming in GinkGo 2.0, but until then we have to structure namespace creation "before each" in a nested format. This will create a number of namespaces for each set of behaviours, such as "this is expected to pass" and "this is expected to fail"