fix(external-dns): add variable structure to Before block
Now we've added [flake-attempts](https://github.com/ministryofjustice/cloud-platform-terraform-concourse/blob/main/pipelines/manager/main/reporting.yaml#L193) to our pipeline, it means failed tests will attempt to rerun. The rerun will use whatever namespace, domain and host are defined in the spec, but this means it'll attempt to use resources that are destroyed in the AfterAll block. This commit adds the namespace, host and domain to the before block, but still make them available to all specs in the suite.